Avoid location.href
It will base the location change off of the `<base href>` URL which is not what we want. If we just change the hash the base URL won't change so the page won't reload. (tested on Chrome)

This does introduce a new step in the browser's history, unfortunately. If this is problematic, the work-around would be to reconstruct the full location URL to bypass the base href.